CI: Refactoring backend tests
