+++ ibis/backends/hdf5/tests/conftest.py
@@ -1,6 +1,23 @@
+from pathlib import Path
+
 import pytest
 
+import ibis
 from ibis.backends.hdf5 import HDFClient
+from ibis.backends.pandas.tests.conftest import PandasTest
+
+
+class HDF5Test(PandasTest):
+    check_names = False
+    supports_divide_by_zero = True
+    returned_timestamp_unit = 'ns'
+
+    @staticmethod
+    def connect(data_directory: Path) -> ibis.client.Client:
+        filename = data_directory / 'functional_alltypes.h5'
+        if not filename.exists():
+            pytest.skip('test data set {} not found'.format(filename))
+        return ibis.hdf5.connect(data_directory)
 
 
 @pytest.fixture
