+++ ibis/backends/pandas/tests/conftest.py
@@ -1,0 +1,35 @@
+from pathlib import Path
+
+import pandas as pd
+
+import ibis
+import ibis.expr.operations as ops
+from ibis.backends.tests.base import BackendTest, RoundHalfToEven
+
+
+class PandasTest(BackendTest, RoundHalfToEven):
+    check_names = False
+    additional_skipped_operations = frozenset({ops.StringSQLLike})
+    supported_to_timestamp_units = BackendTest.supported_to_timestamp_units | {
+        'ns'
+    }
+    supports_divide_by_zero = True
+    returned_timestamp_unit = 'ns'
+
+    @staticmethod
+    def connect(data_directory: Path) -> ibis.client.Client:
+        return ibis.pandas.connect(
+            {
+                'functional_alltypes': pd.read_csv(
+                    str(data_directory / 'functional_alltypes.csv'),
+                    index_col=None,
+                    dtype={'bool_col': bool, 'string_col': str},
+                    parse_dates=['timestamp_col'],
+                    encoding='utf-8',
+                ),
+                'batting': pd.read_csv(str(data_directory / 'batting.csv')),
+                'awards_players': pd.read_csv(
+                    str(data_directory / 'awards_players.csv')
+                ),
+            }
+        )
