+++ ibis/backends/parquet/tests/conftest.py
@@ -1,8 +1,25 @@
+from pathlib import Path
+
 import pyarrow as pa
 import pyarrow.parquet as pq
 import pytest
 
+import ibis
+from ibis.backends.pandas.tests.conftest import PandasTest
 from ibis.backends.parquet import ParquetClient
+
+
+class ParquetTest(PandasTest):
+    check_names = False
+    supports_divide_by_zero = True
+    returned_timestamp_unit = 'ns'
+
+    @staticmethod
+    def connect(data_directory: Path) -> ibis.client.Client:
+        filename = data_directory / 'functional_alltypes.parquet'
+        if not filename.exists():
+            pytest.skip('test data set {} not found'.format(filename))
+        return ibis.parquet.connect(data_directory)
 
 
 @pytest.fixture
