+++ ibis/backends/spark/tests/conftest.py
@@ -2,8 +2,31 @@ import os
 
 import pytest
 
+import ibis.expr.types as ir
 import ibis.util as util
-from ibis.tests.all.conftest import get_spark_testing_client
+from ibis.backends.tests.base import (
+    BackendTest,
+    RoundHalfToEven,
+    get_spark_testing_client,
+)
+
+
+class SparkTest(BackendTest, RoundHalfToEven):
+    @staticmethod
+    def connect(data_directory):
+        return get_spark_testing_client(data_directory)
+
+    @property
+    def functional_alltypes(self) -> ir.TableExpr:
+        return self.connection.table('functional_alltypes')
+
+    @property
+    def batting(self) -> ir.TableExpr:
+        return self.connection.table('batting')
+
+    @property
+    def awards_players(self) -> ir.TableExpr:
+        return self.connection.table('awards_players')
 
 
 @pytest.fixture(scope='session', autouse=True)
