+++ ibis/backends/tests/conftest.py
@@ -1,0 +1,308 @@
+import importlib
+import inspect
+import operator
+import os
+from pathlib import Path
+
+import pytest
+
+import ibis
+import ibis.common.exceptions as com
+import ibis.util as util
+
+from .base import BackendTest
+
+
+def _random_identifier(suffix):
+    return '__ibis_test_{}_{}'.format(suffix, util.guid())
+
+
+def backend_test_classes():
+    """Return all backend tests classes defined in thebackends directories."""
+    for backend_dir in Path(__file__).parent.parent.iterdir():
+        backend = backend_dir.name
+        try:
+            conftest = importlib.import_module(
+                f'ibis.backends.{backend}.tests.conftest'
+            )
+        except ImportError:
+            pass
+        else:
+            for obj_name in dir(conftest):
+                obj = getattr(conftest, obj_name)
+                if (
+                    inspect.isclass(obj)
+                    and issubclass(obj, BackendTest)
+                    and obj.name() == backend
+                ):
+                    yield obj
+
+
+ALL_BACKENDS = sorted(
+    backend_test_classes(), key=operator.attrgetter("__name__")
+)
+
+
+def pytest_runtest_call(item):
+    """Dynamically add various custom markers."""
+    nodeid = item.nodeid
+    for marker in list(item.iter_markers(name="only_on_backends")):
+        (backend_types,) = map(tuple, marker.args)
+        backend = item.funcargs["backend"]
+        assert isinstance(
+            backend, BackendTest
+        ), "backend has type {!r}".format(type(backend).__name__)
+        if not isinstance(backend, backend_types):
+            pytest.skip(
+                f"only_on_backends: {backend} is not in {backend_types} "
+                f"{nodeid}"
+            )
+
+    for marker in list(item.iter_markers(name="skip_backends")):
+        (backend_types,) = map(tuple, marker.args)
+        backend = item.funcargs["backend"]
+        assert isinstance(
+            backend, BackendTest
+        ), "backend has type {!r}".format(type(backend).__name__)
+        if isinstance(backend, backend_types):
+            pytest.skip(f"skip_backends: {backend} {nodeid}")
+
+    for marker in list(item.iter_markers(name="skip_missing_feature")):
+        backend = item.funcargs["backend"]
+        (features,) = marker.args
+        missing_features = [
+            feature for feature in features if not getattr(backend, feature)
+        ]
+        if missing_features:
+            pytest.mark.skip(
+                'Backend {} is missing features {} needed to run {}'.format(
+                    type(backend).__name__, ', '.join(missing_features), nodeid
+                )
+            )
+
+    for marker in list(item.iter_markers(name="xfail_backends")):
+        (backend_types,) = map(tuple, marker.args)
+        backend = item.funcargs["backend"]
+        assert isinstance(
+            backend, BackendTest
+        ), "backend has type {!r}".format(type(backend).__name__)
+        item.add_marker(
+            pytest.mark.xfail(
+                condition=isinstance(backend, backend_types),
+                reason='Backend {} does not pass this test'.format(
+                    type(backend).__name__
+                ),
+                **marker.kwargs,
+            )
+        )
+
+    for marker in list(item.iter_markers(name="xpass_backends")):
+        (backend_types,) = map(tuple, marker.args)
+        backend = item.funcargs["backend"]
+        assert isinstance(
+            backend, BackendTest
+        ), "backend has type {!r}".format(type(backend).__name__)
+        item.add_marker(
+            pytest.mark.xfail(
+                condition=not isinstance(backend, backend_types),
+                reason='{} does not pass this test'.format(
+                    type(backend).__name__
+                ),
+                **marker.kwargs,
+            )
+        )
+
+
+@pytest.hookimpl(hookwrapper=True)
+def pytest_pyfunc_call(pyfuncitem):
+    """Dynamically add an xfail marker for specific backends."""
+    outcome = yield
+    try:
+        outcome.get_result()
+    except (
+        com.OperationNotDefinedError,
+        com.UnsupportedOperationError,
+        com.UnsupportedBackendType,
+        NotImplementedError,
+    ) as e:
+        markers = list(pyfuncitem.iter_markers(name="xfail_unsupported"))
+        assert (
+            len(markers) == 1
+        ), "More than one xfail_unsupported marker found on test {}".format(
+            pyfuncitem
+        )
+        (marker,) = markers
+        backend = pyfuncitem.funcargs["backend"]
+        assert isinstance(
+            backend, BackendTest
+        ), "backend has type {!r}".format(type(backend).__name__)
+        pytest.xfail(reason='{}: {}'.format(type(backend).__name__, e))
+
+
+pytestmark = pytest.mark.backend
+
+pytest_backends = os.environ.get('PYTEST_BACKENDS', '').split(' ')
+params_backend = [
+    pytest.param(backend, marks=getattr(pytest.mark, backend.name()))
+    for backend in ALL_BACKENDS
+    if backend.name() in pytest_backends or not pytest_backends
+]
+if len(pytest_backends) != len(params_backend):
+    unknown_backends = set(pytest_backends) - {b.name() for b in ALL_BACKENDS}
+    raise ValueError(
+        'PYTEST_BACKENDS environment variable contains unknown '
+        f'backends {unknown_backends} {[b.name() for b in ALL_BACKENDS]}'
+    )
+
+
+@pytest.fixture(params=params_backend, scope='session')
+def backend(request, data_directory):
+    return request.param(data_directory)
+
+
+@pytest.fixture(scope='session')
+def con(backend):
+    return backend.connection
+
+
+@pytest.fixture(scope='session')
+def alltypes(backend):
+    return backend.functional_alltypes
+
+
+@pytest.fixture(scope='session')
+def sorted_alltypes(alltypes):
+    return alltypes.sort_by('id')
+
+
+@pytest.fixture(scope='session')
+def batting(backend):
+    return backend.batting
+
+
+@pytest.fixture(scope='session')
+def awards_players(backend):
+    return backend.awards_players
+
+
+@pytest.fixture(scope='session')
+def geo(backend):
+    if backend.geo is None:
+        pytest.skip(
+            'Geo Spatial type not supported for {} backend.'.format(
+                backend.name
+            )
+        )
+    return backend.geo
+
+
+@pytest.fixture
+def analytic_alltypes(alltypes):
+    return alltypes
+
+
+@pytest.fixture(scope='session')
+def df(alltypes):
+    return alltypes.execute()
+
+
+@pytest.fixture(scope='session')
+def sorted_df(df):
+    return df.sort_values('id').reset_index(drop=True)
+
+
+@pytest.fixture(scope='session')
+def batting_df(batting):
+    return batting.execute(limit=None)
+
+
+@pytest.fixture(scope='session')
+def awards_players_df(awards_players):
+    return awards_players.execute(limit=None)
+
+
+@pytest.fixture(scope='session')
+def geo_df(geo):
+    # Currently geo is implemented just for OmniSciDB
+    if geo is not None:
+        return geo.execute(limit=None)
+    return None
+
+
+@pytest.fixture
+def temp_table(con: ibis.client.Client) -> str:
+    """
+    Return a temporary table name.
+
+    Parameters
+    ----------
+    con : ibis.client.Client
+
+    Yields
+    ------
+    name : string
+        Random table name for a temporary usage.
+    """
+    name = _random_identifier('table')
+    try:
+        yield name
+    finally:
+        if hasattr(con, 'drop_table'):
+            con.drop_table(name, force=True)
+
+
+@pytest.fixture
+def temp_view(con) -> str:
+    """Return a temporary view name.
+
+    Parameters
+    ----------
+    con : ibis.omniscidb.OmniSciDBClient
+
+    Yields
+    ------
+    name : string
+        Random view name for a temporary usage.
+    """
+    name = _random_identifier('view')
+    try:
+        yield name
+    finally:
+        if hasattr(con, 'drop_view'):
+            con.drop_view(name, force=True)
+
+
+@pytest.fixture(scope='session')
+def current_data_db(con, backend) -> str:
+    """Return current database name."""
+    if not hasattr(con, 'current_database'):
+        pytest.skip(
+            f'{backend.name} backend doesn\'t have current_database method.'
+        )
+    return con.current_database
+
+
+@pytest.fixture
+def alternate_current_database(con, backend, current_data_db: str) -> str:
+    """Create a temporary database and yield its name.
+    Drops the created database upon completion.
+
+    Parameters
+    ----------
+    con : ibis.client.Client
+    current_data_db : str
+    Yields
+    -------
+    str
+    """
+    name = _random_identifier('database')
+    if not hasattr(con, 'create_database'):
+        pytest.skip(
+            f'{backend.name} backend doesn\'t have create_database method.'
+        )
+    con.create_database(name)
+    try:
+        yield name
+    finally:
+        con.set_database(current_data_db)
+        con.drop_database(name, force=True)
