+++ ibis/backends/tests/test_join.py
@@ -2,10 +2,12 @@ import pandas as pd
 import pytest
 from pytest import param
 
-from ibis.tests.backends import Csv, Pandas, PySpark
+from ibis.backends.csv.tests.conftest import CsvTest
+from ibis.backends.pandas.tests.conftest import PandasTest
+from ibis.backends.pyspark.tests.conftest import PySparkTest
 
 # add here backends that passes join tests
-all_db_join_supported = [Pandas, PySpark]
+all_db_join_supported = [PandasTest, PySparkTest]
 
 
 @pytest.mark.parametrize(
@@ -31,7 +33,7 @@ all_db_join_supported = [Pandas, PySpark]
 )
 @pytest.mark.only_on_backends(all_db_join_supported)
 # Csv is a subclass of Pandas so need to skip it explicitly.
-@pytest.mark.skip_backends([Csv])
+@pytest.mark.skip_backends([CsvTest])
 @pytest.mark.xfail_unsupported
 def test_join_project_left_table(backend, con, batting, awards_players, how):
