+++ ibis/backends/tests/test_json.py
@@ -5,10 +5,10 @@ import pytest
 from pytest import param
 
 import ibis
-from ibis.tests.backends import Postgres
+from ibis.backends.postgres.tests.conftest import PostgresTest
 
 # add here backends that support json types
-all_db_geo_supported = [Postgres]
+all_db_geo_supported = [PostgresTest]
 
 
 @pytest.mark.parametrize('data', [param({'status': True}, id='status')])
