+++ ibis/backends/tests/test_union.py
@@ -1,11 +1,17 @@
 import pandas as pd
 import pytest
 
-from ibis.tests.backends import BigQuery, Impala, Pandas, Postgres, PySpark
+from ibis.backends.bigquery.tests.conftest import BigQueryTest
+from ibis.backends.impala.tests.conftest import ImpalaTest
+from ibis.backends.pandas.tests.conftest import PandasTest
+from ibis.backends.postgres.tests.conftest import PostgresTest
+from ibis.backends.pyspark.tests.conftest import PySparkTest
 
 
 @pytest.mark.parametrize('distinct', [False, True])
-@pytest.mark.only_on_backends([BigQuery, Impala, Pandas, Postgres, PySpark])
+@pytest.mark.only_on_backends(
+    [BigQueryTest, ImpalaTest, PandasTest, PostgresTest, PySparkTest]
+)
 @pytest.mark.xfail_unsupported
 def test_union(backend, alltypes, df, distinct):
     result = alltypes.union(alltypes, distinct=distinct).execute()
